feat: integrate widget into blog
